pyproject.toml
[project]
name = "transcribe-audio-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20240930",
]
